refactored handlers, change amount and points type from int32 to uint64

+++ internal/assets/migrations/001_initial.sql
@@ -6,7 +6,7 @@ AS $$ BEGIN NEW.updated_at = EXTRACT('EPOCH' FROM NOW()); RETURN NEW; END; $$;
 CREATE TABLE IF NOT EXISTS balances
 (
     did              text PRIMARY KEY,
-    amount           integer not null default 0,
+    amount           bigint not null default 0,
     created_at       integer not null default EXTRACT('EPOCH' FROM NOW()),
     updated_at       integer not null default EXTRACT('EPOCH' FROM NOW()),
     passport_hash    text UNIQUE,
