+++ resources/model_withdrawal_attributes.go
@@ -8,7 +8,7 @@ type WithdrawalAttributes struct {
 	// Rarimo address which points were withdrawn to. Can be any valid address.
 	Address string `json:"address"`
 	// Amount of points withdrawn
-	Amount int32 `json:"amount"`
+	Amount uint64 `json:"amount"`
 	// Unix timestamp of withdrawal creation
 	CreatedAt int32 `json:"created_at"`
 }
